Move RoleDataMapping out of autogenerated folder
