+++ clients/admin-ui/src/types/api/index.ts
@@ -200,7 +200,6 @@ export type { ResourceFilter } from "./models/ResourceFilter";
 export { ResourceTypes } from "./models/ResourceTypes";
 export { ResponseFormat } from "./models/ResponseFormat";
 export type { ReviewPrivacyRequestIds } from "./models/ReviewPrivacyRequestIds";
-export type { ROLES } from "./models/RolesDataMapping";
 export { RoleRegistryEnum } from "./models/RoleRegistryEnum";
 export type { RowCountRequest } from "./models/RowCountRequest";
 export type { RuleCreate } from "./models/RuleCreate";
